refactor: tester + test pattern extract
