+++ packages/amplify-graphql-api-construct-tests/src/sql-tests-common/sql-oidc-auth.ts
@@ -1,33 +1,13 @@
 import { ImportedRDSType } from '@aws-amplify/graphql-transformer-core';
-import { AUTH_TYPE } from 'aws-appsync';
-import { gql } from 'graphql-tag';
-import { schema as generateSchema } from './tests-sources/sql-oidc-auth/provider';
-import { getUserMap } from '../utils/sql-cognito-helper';
-import { configureAppSyncClients } from '../utils/sql-appsync-client-helper';
-import { loggedInUserCrudOperationsTest } from './sql-auth-tests';
-import {
-  createModelOperationHelpers,
-  checkOperationResult,
-  checkListItemExistence,
-} from '../utils/appsync-model-operation/model-operation-helper';
-import { TestOptions, setupTest, cleanupTest } from '../utils/sql-test-config-helper';
-import { stackConfig as generateStackConfig } from './tests-sources/sql-oidc-auth/stack-config';
+import { DynamicModelAuthTester } from './tests-sources/sql-dynamic-model-auth/tester';
+import { TestConfigOutput, TestOptions, setupTest, cleanupTest } from '../utils/sql-test-config-helper';
+import { stackConfig as generateStackConfig } from './tests-sources//sql-dynamic-model-auth/sql-oidc-auth/stack-config';
 import { authConstructDependency } from '../__tests__/additional-dependencies';
 
 export const testGraphQLAPIWithOIDCAccess = (options: TestOptions, testBlockDescription: string, engine: ImportedRDSType): void => {
   describe(`${testBlockDescription} - ${engine}`, () => {
-    const schema = generateSchema(engine);
-    const authProvider = AUTH_TYPE.OPENID_CONNECT;
-
-    const userName1 = 'user1@amazon.com';
-    const userName2 = 'user2@amazon.com';
-    const password = 'Password1234!';
-    const adminGroupName = 'Admin';
-    const devGroupName = 'Dev';
-
-    let userMap = {};
-    let appSyncClients = {};
-    let testConfigOutput;
+    let testConfigOutput: TestConfigOutput;
+    let authTester: DynamicModelAuthTester;
 
     beforeAll(async () => {
       testConfigOutput = await setupTest({
@@ -36,15 +16,8 @@ export const testGraphQLAPIWithOIDCAccess = (options: TestOptions, testBlockDesc
         additionalDependencies: [authConstructDependency],
       });
 
-      // const cognitoIdentityPoolCredentialsManager = new CognitoUserPoolAuthHelper(testConfigOutput as UserPoolAuthConstructStackOutputs);
-      // await cognitoIdentityPoolCredentialsManager.createUser({ username: userName1, email: userName1, password }, [adminGroupName]);
-      // await cognitoIdentityPoolCredentialsManager.createUser({ username: userName2, email: userName2, password }, [devGroupName]);
-
-      // userMap[userName1] = await cognitoIdentityPoolCredentialsManager.getAuthRoleCredentials({ username: userName1, password });
-      // userMap[userName2] = await cognitoIdentityPoolCredentialsManager.getAuthRoleCredentials({ username: userName2, password });
-
-      userMap = await getUserMap(testConfigOutput);
-      appSyncClients = await configureAppSyncClients(testConfigOutput, userMap);
+      authTester = new DynamicModelAuthTester(testConfigOutput);
+      await authTester.initialize();
     });
 
     afterAll(async () => {
@@ -52,726 +25,75 @@ export const testGraphQLAPIWithOIDCAccess = (options: TestOptions, testBlockDesc
     });
 
     test('logged in user can perform CRUD and subscription operations', async () => {
-      // const modelName = 'TodoPrivate';
-      // const modelOperationHelpers = createModelOperationHelpers(appSyncClients[authProvider][userName1], schema);
-      // const todoHelper = modelOperationHelpers[modelName];
-
-      // const todo = {
-      //   content: 'Todo',
-      // };
-      // const resultSetName = `create${modelName}`;
-      // const createResult = await todoHelper.create(resultSetName, todo);
-      // expect(createResult.data[resultSetName].id).toBeDefined();
-      // todo['id'] = createResult.data[resultSetName].id;
-      // checkOperationResult(createResult, todo, resultSetName);
-
-      // const todoUpdated = {
-      //   id: todo['id'],
-      //   content: 'Todo updated',
-      // };
-      // const updateResult = await todoHelper.update(`update${modelName}`, todoUpdated);
-      // checkOperationResult(updateResult, todoUpdated, `update${modelName}`);
-
-      // const getResult = await todoHelper.get({
-      //   id: todo['id'],
-      // });
-      // checkOperationResult(getResult, todoUpdated, `get${modelName}`);
-
-      // const listTodosResult = await todoHelper.list();
-      // checkListItemExistence(listTodosResult, `list${modelName}s`, todo['id'], true);
-
-      // const deleteResult = await todoHelper.delete(`delete${modelName}`, {
-      //   id: todo['id'],
-      // });
-      // checkOperationResult(deleteResult, todoUpdated, `delete${modelName}`);
-
-      await loggedInUserCrudOperationsTest({
-        appSyncClients,
-        userMap,
-        authProvider,
-        schema,
-      });
+      await authTester.testLoggedInUserCrudOperations();
     });
 
-    // test('owner of a record can perform CRUD and subscription operations using default owner field', async () => {
-    //   const modelName = 'TodoOwner';
-    //   const modelOperationHelpers = createModelOperationHelpers(appSyncClients[authProvider][userName1], schema);
-    //   const todoHelper = modelOperationHelpers[modelName];
-
-    //   const todo = {
-    //     content: 'Todo',
-    //   };
-    //   const resultSetName = `create${modelName}`;
-    //   const createResult = await todoHelper.create(resultSetName, todo);
-    //   expect(createResult.data[resultSetName].id).toBeDefined();
-    //   todo['id'] = createResult.data[resultSetName].id;
-    //   expect(createResult.data[resultSetName].content).toEqual(todo.content);
-    //   expect(createResult.data[resultSetName].owner).toBeDefined();
-
-    //   const todoWithOwner = {
-    //     ...todo,
-    //     owner: createResult.data[resultSetName].owner,
-    //   };
-
-    //   const todoUpdated = {
-    //     id: todo['id'],
-    //     content: 'Todo updated',
-    //     owner: todoWithOwner.owner,
-    //   };
-    //   const updateResult = await todoHelper.update(`update${modelName}`, todoUpdated);
-    //   checkOperationResult(updateResult, todoUpdated, `update${modelName}`);
-
-    //   const getResult = await todoHelper.get({
-    //     id: todo['id'],
-    //   });
-    //   checkOperationResult(getResult, todoUpdated, `get${modelName}`);
-
-    //   const listTodosResult = await todoHelper.list();
-    //   checkListItemExistence(listTodosResult, `list${modelName}s`, todo['id'], true);
-
-    //   const deleteResult = await todoHelper.delete(`delete${modelName}`, {
-    //     id: todo['id'],
-    //   });
-    //   checkOperationResult(deleteResult, todoUpdated, `delete${modelName}`);
-    // });
-
-    // test('non-owner of a record cannot access or subscribe to it using default owner field', async () => {
-    //   const modelName = 'TodoOwner';
-    //   const modelOperationHelpersOwner = createModelOperationHelpers(appSyncClients[authProvider][userName1], schema);
-    //   const modelOperationHelpersNonOwner = createModelOperationHelpers(appSyncClients[authProvider][userName2], schema);
-    //   const todoHelperOwner = modelOperationHelpersOwner[modelName];
-    //   const todoHelperNonOwner = modelOperationHelpersNonOwner[modelName];
-
-    //   const todo = {
-    //     content: 'Todo',
-    //   };
-    //   const resultSetName = `create${modelName}`;
-    //   const createResult = await todoHelperOwner.create(resultSetName, todo);
-    //   expect(createResult.data[resultSetName].id).toBeDefined();
-    //   todo['id'] = createResult.data[resultSetName].id;
-
-    //   const todoUpdated = {
-    //     id: todo['id'],
-    //     content: 'Todo updated',
-    //     owner: userName2,
-    //   };
-    //   await expect(todoHelperNonOwner.update(`update${modelName}`, todoUpdated)).rejects.toThrow(
-    //     'GraphQL error: Not Authorized to access updateTodoOwner on type Mutation',
-    //   );
-
-    //   await expect(async () => {
-    //     const getResult = await todoHelperNonOwner.get({
-    //       id: todo['id'],
-    //     });
-    //   }).rejects.toThrowErrorMatchingInlineSnapshot(`"GraphQL error: Not Authorized to access getTodoOwner on type Query"`);
-
-    //   const listTodosResult = await todoHelperNonOwner.list();
-    //   checkListItemExistence(listTodosResult, `list${modelName}s`, todo['id']);
-
-    //   await expect(
-    //     todoHelperNonOwner.delete(`delete${modelName}`, {
-    //       id: todo['id'],
-    //     }),
-    //   ).rejects.toThrow('GraphQL error: Not Authorized to access deleteTodoOwner on type Mutation');
-    // });
-
-    // test('custom owner field used to store owner information', async () => {
-    //   const modelName = 'TodoOwnerFieldString';
-    //   const modelOperationHelpers = createModelOperationHelpers(appSyncClients[authProvider][userName1], schema);
-    //   const todoHelper = modelOperationHelpers[modelName];
-
-    //   const todo = {
-    //     content: 'Todo',
-    //   };
-    //   const resultSetName = `create${modelName}`;
-    //   const createResult = await todoHelper.create(resultSetName, todo);
-    //   expect(createResult.data[resultSetName].id).toBeDefined();
-    //   todo['id'] = createResult.data[resultSetName].id;
-    //   expect(createResult.data[resultSetName].content).toEqual(todo.content);
-    //   expect(createResult.data[resultSetName].author).toEqual(userName1);
-
-    //   const todoWithOwner = {
-    //     ...todo,
-    //     author: userName1,
-    //   };
-
-    //   const todo1Updated = {
-    //     id: todo['id'],
-    //     content: 'Todo updated',
-    //     author: todoWithOwner.author,
-    //   };
-    //   const updateResult = await todoHelper.update(`update${modelName}`, todo1Updated);
-    //   checkOperationResult(updateResult, todo1Updated, `update${modelName}`);
-
-    //   const getResult = await todoHelper.get({
-    //     id: todo['id'],
-    //   });
-    //   checkOperationResult(getResult, todo1Updated, `get${modelName}`);
-
-    //   const listTodosResult = await todoHelper.list();
-    //   checkListItemExistence(listTodosResult, `list${modelName}s`, todo['id'], true);
-
-    //   const deleteResult = await todoHelper.delete(`delete${modelName}`, {
-    //     id: todo['id'],
-    //   });
-    //   checkOperationResult(deleteResult, todo1Updated, `delete${modelName}`);
-    // });
-
-    // test('non-owner of a record cannot pretend to be an owner and gain access', async () => {
-    //   const modelName = 'TodoOwnerFieldString';
-    //   const modelOperationHelpersOwner = createModelOperationHelpers(appSyncClients[authProvider][userName1], schema);
-    //   const modelOperationHelpersNonOwner = createModelOperationHelpers(appSyncClients[authProvider][userName2], schema);
-    //   const todoHelperOwner = modelOperationHelpersOwner[modelName];
-    //   const todoHelperNonOwner = modelOperationHelpersNonOwner[modelName];
-
-    //   const todo = {
-    //     content: 'Todo',
-    //   };
-    //   const resultSetName = `create${modelName}`;
-    //   const createResult = await todoHelperOwner.create(resultSetName, todo);
-    //   expect(createResult.data[resultSetName].id).toBeDefined();
-    //   todo['id'] = createResult.data[resultSetName].id;
-
-    //   const todoUpdated = {
-    //     id: todo['id'],
-    //     content: 'Todo updated',
-    //     author: userName1,
-    //   };
-    //   await expect(todoHelperNonOwner.update(`update${modelName}`, todoUpdated)).rejects.toThrow(
-    //     'GraphQL error: Not Authorized to access updateTodoOwnerFieldString on type Mutation',
-    //   );
-
-    //   await expect(async () => {
-    //     const getResult = await todoHelperNonOwner.get({
-    //       id: todo['id'],
-    //     });
-    //   }).rejects.toThrowErrorMatchingInlineSnapshot(`"GraphQL error: Not Authorized to access getTodoOwnerFieldString on type Query"`);
-
-    //   const listTodosResult = await todoHelperNonOwner.list();
-    //   checkListItemExistence(listTodosResult, `list${modelName}s`, todo['id']);
-
-    //   await expect(
-    //     todoHelperNonOwner.delete(`delete${modelName}`, {
-    //       id: todo['id'],
-    //     }),
-    //   ).rejects.toThrow('GraphQL error: Not Authorized to access deleteTodoOwnerFieldString on type Mutation');
-    // });
-
-    // test('member in list of owners can perform CRUD and subscription operations', async () => {
-    //   const modelName = 'TodoOwnerFieldList';
-    //   const modelOperationHelpers = createModelOperationHelpers(appSyncClients[authProvider][userName1], schema);
-    //   const todoHelper = modelOperationHelpers[modelName];
-
-    //   const todo = {
-    //     content: 'Todo',
-    //     authors: [userName1],
-    //   };
-    //   const resultSetName = `create${modelName}`;
-    //   const createResult = await todoHelper.create(resultSetName, todo);
-    //   expect(createResult.data[resultSetName].id).toBeDefined();
-    //   todo['id'] = createResult.data[resultSetName].id;
-    //   expect(createResult.data[resultSetName].content).toEqual(todo.content);
-    //   expect(createResult.data[resultSetName].authors).toEqual([userName1]);
-
-    //   const todoUpdated = {
-    //     id: todo['id'],
-    //     content: 'Todo updated',
-    //     authors: [userName1],
-    //   };
-    //   const updateResult = await todoHelper.update(`update${modelName}`, todoUpdated);
-    //   checkOperationResult(updateResult, todoUpdated, `update${modelName}`);
-
-    //   const getResult = await todoHelper.get({
-    //     id: todo['id'],
-    //   });
-    //   checkOperationResult(getResult, todoUpdated, `get${modelName}`);
-
-    //   const listTodosResult = await todoHelper.list();
-    //   checkListItemExistence(listTodosResult, `list${modelName}s`, todo['id'], true);
-
-    //   const deleteResult = await todoHelper.delete(`delete${modelName}`, {
-    //     id: todo['id'],
-    //   });
-    //   checkOperationResult(deleteResult, todoUpdated, `delete${modelName}`);
-    // });
-
-    // test('non-owner of a record cannot add themself to owner list', async () => {
-    //   const modelName = 'TodoOwnerFieldList';
-    //   const modelOperationHelpersOwner = createModelOperationHelpers(appSyncClients[authProvider][userName1], schema);
-    //   const modelOperationHelpersNonOwner = createModelOperationHelpers(appSyncClients[authProvider][userName2], schema);
-    //   const todoHelperOwner = modelOperationHelpersOwner[modelName];
-    //   const todoHelperNonOwner = modelOperationHelpersNonOwner[modelName];
-
-    //   const todo = {
-    //     content: 'Todo',
-    //     authors: [userName1],
-    //   };
-    //   const resultSetName = `create${modelName}`;
-    //   const createResult = await todoHelperOwner.create(resultSetName, todo);
-    //   expect(createResult.data[resultSetName].id).toBeDefined();
-    //   todo['id'] = createResult.data[resultSetName].id;
-
-    //   const todoUpdated = {
-    //     id: todo['id'],
-    //     content: 'Todo updated',
-    //     authors: [userName1, userName2],
-    //   };
-    //   await expect(todoHelperNonOwner.update(`update${modelName}`, todoUpdated)).rejects.toThrow(
-    //     'GraphQL error: Not Authorized to access updateTodoOwnerFieldList on type Mutation',
-    //   );
-
-    //   await expect(async () => {
-    //     const getResult = await todoHelperNonOwner.get({
-    //       id: todo['id'],
-    //     });
-    //   }).rejects.toThrowErrorMatchingInlineSnapshot(`"GraphQL error: Not Authorized to access getTodoOwnerFieldList on type Query"`);
-
-    //   const listTodosResult = await todoHelperNonOwner.list();
-    //   checkListItemExistence(listTodosResult, `list${modelName}s`, todo['id']);
-
-    //   await expect(
-    //     todoHelperNonOwner.delete(`delete${modelName}`, {
-    //       id: todo['id'],
-    //     }),
-    //   ).rejects.toThrow('GraphQL error: Not Authorized to access deleteTodoOwnerFieldList on type Mutation');
-    // });
-
-    // test('owner can add another user to the owner list', async () => {
-    //   const modelName = 'TodoOwnerFieldList';
-    //   const modelOperationHelpersOwner = createModelOperationHelpers(appSyncClients[authProvider][userName1], schema);
-    //   const modelOperationHelpersNonOwner = createModelOperationHelpers(appSyncClients[authProvider][userName2], schema);
-    //   const todoHelperOwner = modelOperationHelpersOwner[modelName];
-    //   const todoHelperAnotherOwner = modelOperationHelpersNonOwner[modelName];
-
-    //   const todo = {
-    //     content: 'Todo',
-    //     authors: [userName1, userName2],
-    //   };
-    //   const resultSetName = `create${modelName}`;
-    //   const createResult = await todoHelperOwner.create(resultSetName, todo);
-    //   expect(createResult.data[resultSetName].id).toBeDefined();
-    //   todo['id'] = createResult.data[resultSetName].id;
-
-    //   const todoUpdated = {
-    //     id: todo['id'],
-    //     content: 'Todo updated',
-    //     authors: [userName1, userName2],
-    //   };
-    //   const updateResult = await todoHelperAnotherOwner.update(`update${modelName}`, todoUpdated);
-    //   checkOperationResult(updateResult, todoUpdated, `update${modelName}`);
-
-    //   const getResult = await todoHelperAnotherOwner.get({
-    //     id: todo['id'],
-    //   });
-    //   checkOperationResult(getResult, todoUpdated, `get${modelName}`);
-
-    //   const listTodosResult = await todoHelperAnotherOwner.list();
-    //   checkListItemExistence(listTodosResult, `list${modelName}s`, todo['id'], true);
-
-    //   const deleteResult = await todoHelperAnotherOwner.delete(`delete${modelName}`, {
-    //     id: todo['id'],
-    //   });
-    //   checkOperationResult(deleteResult, todoUpdated, `delete${modelName}`);
-    // });
-
-    // test('users in static group can perform CRUD and subscription operations', async () => {
-    //   const modelName = 'TodoStaticGroup';
-    //   const modelOperationHelpers = createModelOperationHelpers(appSyncClients[authProvider][userName1], schema);
-    //   const todoHelper = modelOperationHelpers[modelName];
-
-    //   const todo = {
-    //     content: 'Todo',
-    //   };
-    //   const resultSetName = `create${modelName}`;
-    //   const createResult = await todoHelper.create(resultSetName, todo);
-    //   expect(createResult.data[resultSetName].id).toBeDefined();
-    //   todo['id'] = createResult.data[resultSetName].id;
-    //   expect(createResult.data[resultSetName].content).toEqual(todo.content);
-
-    //   const todoUpdated = {
-    //     id: todo['id'],
-    //     content: 'Todo updated',
-    //   };
-    //   const updateResult = await todoHelper.update(`update${modelName}`, todoUpdated);
-    //   checkOperationResult(updateResult, todoUpdated, `update${modelName}`);
-
-    //   const getResult = await todoHelper.get({
-    //     id: todo['id'],
-    //   });
-    //   checkOperationResult(getResult, todoUpdated, `get${modelName}`);
-
-    //   const listTodosResult = await todoHelper.list();
-    //   checkListItemExistence(listTodosResult, `list${modelName}s`, todo['id'], true);
-
-    //   const deleteResult = await todoHelper.delete(`delete${modelName}`, {
-    //     id: todo['id'],
-    //   });
-    //   checkOperationResult(deleteResult, todoUpdated, `delete${modelName}`);
-    // });
-
-    // test('users not in static group cannot perform CRUD operations', async () => {
-    //   const modelName = 'TodoStaticGroup';
-    //   const modelOperationHelpersAdmin = createModelOperationHelpers(appSyncClients[authProvider][userName1], schema);
-    //   const modelOperationHelpersNonAdmin = createModelOperationHelpers(appSyncClients[authProvider][userName2], schema);
-    //   const todoHelperAdmin = modelOperationHelpersAdmin[modelName];
-    //   const todoHelperNonAdmin = modelOperationHelpersNonAdmin[modelName];
-
-    //   const todo = {
-    //     content: 'Todo',
-    //   };
-    //   const resultSetName = `create${modelName}`;
-    //   const createResult = await todoHelperAdmin.create(resultSetName, todo);
-    //   expect(createResult.data[resultSetName].id).toBeDefined();
-    //   todo['id'] = createResult.data[resultSetName].id;
-
-    //   await expect(todoHelperNonAdmin.create(`create${modelName}`, todo)).rejects.toThrow(
-    //     'GraphQL error: Not Authorized to access createTodoStaticGroup on type Mutation',
-    //   );
-
-    //   const todoUpdated = {
-    //     id: todo['id'],
-    //     content: 'Todo updated',
-    //   };
-    //   await expect(todoHelperNonAdmin.update(`update${modelName}`, todoUpdated)).rejects.toThrow(
-    //     'GraphQL error: Not Authorized to access updateTodoStaticGroup on type Mutation',
-    //   );
-
-    //   expect(
-    //     todoHelperNonAdmin.get({
-    //       id: todo['id'],
-    //     }),
-    //   ).rejects.toThrow('GraphQL error: Not Authorized to access getTodoStaticGroup on type Query');
-
-    //   await expect(todoHelperNonAdmin.list()).rejects.toThrow('GraphQL error: Not Authorized to access listTodoStaticGroups on type Query');
-
-    //   await expect(
-    //     todoHelperNonAdmin.delete(`delete${modelName}`, {
-    //       id: todo['id'],
-    //     }),
-    //   ).rejects.toThrow('GraphQL error: Not Authorized to access deleteTodoStaticGroup on type Mutation');
-    // });
-
-    // test('users in group stored as string can perform CRUD and subscription operations', async () => {
-    //   const modelName = 'TodoGroupFieldString';
-    //   const modelOperationHelpers = createModelOperationHelpers(appSyncClients[authProvider][userName1], schema);
-    //   const todoHelper = modelOperationHelpers[modelName];
-
-    //   const todo = {
-    //     content: 'Todo',
-    //     groupField: adminGroupName,
-    //   };
-    //   const resultSetName = `create${modelName}`;
-    //   const createResult = await todoHelper.create(resultSetName, todo);
-    //   expect(createResult.data[resultSetName].id).toBeDefined();
-    //   todo['id'] = createResult.data[resultSetName].id;
-    //   expect(createResult.data[resultSetName].content).toEqual(todo.content);
-    //   expect(createResult.data[resultSetName].groupField).toEqual(adminGroupName);
-
-    //   const todoUpdated = {
-    //     id: todo['id'],
-    //     content: 'Todo updated',
-    //     groupField: adminGroupName,
-    //   };
-    //   const updateResult = await todoHelper.update(`update${modelName}`, todoUpdated);
-    //   checkOperationResult(updateResult, todoUpdated, `update${modelName}`);
-
-    //   const getResult = await todoHelper.get({
-    //     id: todo['id'],
-    //   });
-    //   checkOperationResult(getResult, todoUpdated, `get${modelName}`);
-
-    //   const listTodosResult = await todoHelper.list();
-    //   checkListItemExistence(listTodosResult, `list${modelName}s`, todo['id'], true);
-
-    //   const deleteResult = await todoHelper.delete(`delete${modelName}`, {
-    //     id: todo['id'],
-    //   });
-    //   checkOperationResult(deleteResult, todoUpdated, `delete${modelName}`);
-    // });
-
-    // test('users cannot spoof their group membership and gain access', async () => {
-    //   const modelName = 'TodoGroupFieldString';
-    //   const modelOperationHelpersAdmin = createModelOperationHelpers(appSyncClients[authProvider][userName1], schema);
-    //   const modelOperationHelpersNonAdmin = createModelOperationHelpers(appSyncClients[authProvider][userName2], schema);
-    //   const todoHelperAdmin = modelOperationHelpersAdmin[modelName];
-    //   const todoHelperNonAdmin = modelOperationHelpersNonAdmin[modelName];
-
-    //   const todo = {
-    //     content: 'Todo',
-    //     groupField: adminGroupName,
-    //   };
-    //   const resultSetName = `create${modelName}`;
-    //   const createResult = await todoHelperAdmin.create(resultSetName, todo);
-    //   expect(createResult.data[resultSetName].id).toBeDefined();
-    //   todo['id'] = createResult.data[resultSetName].id;
-
-    //   await expect(todoHelperNonAdmin.create(resultSetName, todo)).rejects.toThrow(
-    //     'GraphQL error: Not Authorized to access createTodoGroupFieldString on type Mutation',
-    //   );
-
-    //   const todoUpdated = {
-    //     id: todo['id'],
-    //     content: 'Todo updated',
-    //     groupField: devGroupName,
-    //   };
-    //   await expect(todoHelperNonAdmin.update(`update${modelName}`, todoUpdated)).rejects.toThrow(
-    //     'GraphQL error: Not Authorized to access updateTodoGroupFieldString on type Mutation',
-    //   );
-
-    //   await expect(async () => {
-    //     const getResult = await todoHelperNonAdmin.get({
-    //       id: todo['id'],
-    //     });
-    //   }).rejects.toThrowErrorMatchingInlineSnapshot(`"GraphQL error: Not Authorized to access getTodoGroupFieldString on type Query"`);
-
-    //   const listTodosResult = await todoHelperNonAdmin.list();
-    //   checkListItemExistence(listTodosResult, `list${modelName}s`, todo['id']);
-
-    //   await expect(
-    //     todoHelperNonAdmin.delete(`delete${modelName}`, {
-    //       id: todo['id'],
-    //     }),
-    //   ).rejects.toThrow('GraphQL error: Not Authorized to access deleteTodoGroupFieldString on type Mutation');
-    // });
-
-    // test('users in groups stored as list can perform CRUD and subscription operations', async () => {
-    //   const modelName = 'TodoGroupFieldList';
-    //   const modelOperationHelpers = createModelOperationHelpers(appSyncClients[authProvider][userName1], schema);
-    //   const todoHelper = modelOperationHelpers[modelName];
-
-    //   const todo = {
-    //     content: 'Todo',
-    //     groupsField: [adminGroupName],
-    //   };
-    //   const resultSetName = `create${modelName}`;
-    //   const createResult = await todoHelper.create(resultSetName, todo);
-    //   expect(createResult.data[resultSetName].id).toBeDefined();
-    //   todo['id'] = createResult.data[resultSetName].id;
-    //   expect(createResult.data[resultSetName].content).toEqual(todo.content);
-    //   expect(createResult.data[resultSetName].groupsField).toEqual([adminGroupName]);
-
-    //   const todo1Updated = {
-    //     id: todo['id'],
-    //     content: 'Todo updated',
-    //     groupsField: [adminGroupName],
-    //   };
-    //   const updateResult = await todoHelper.update(`update${modelName}`, todo1Updated);
-    //   checkOperationResult(updateResult, todo1Updated, `update${modelName}`);
-
-    //   const getResult = await todoHelper.get({
-    //     id: todo['id'],
-    //   });
-    //   checkOperationResult(getResult, todo1Updated, `get${modelName}`);
-
-    //   const listTodosResult = await todoHelper.list();
-    //   checkListItemExistence(listTodosResult, `list${modelName}s`, todo['id'], true);
-
-    //   const deleteResult = await todoHelper.delete(`delete${modelName}`, {
-    //     id: todo['id'],
-    //   });
-    //   checkOperationResult(deleteResult, todo1Updated, `delete${modelName}`);
-    // });
-
-    // test('users not part of allowed groups cannot access the records or modify allowed groups', async () => {
-    //   const modelName = 'TodoGroupFieldList';
-    //   const modelOperationHelpersAdmin = createModelOperationHelpers(appSyncClients[authProvider][userName1], schema);
-    //   const modelOperationHelpersNonAdmin = createModelOperationHelpers(appSyncClients[authProvider][userName2], schema);
-    //   const todoHelperAdmin = modelOperationHelpersAdmin[modelName];
-    //   const todoHelperNonAdmin = modelOperationHelpersNonAdmin[modelName];
-
-    //   const todo = {
-    //     content: 'Todo',
-    //     groupsField: [adminGroupName],
-    //   };
-    //   const resultSetName = `create${modelName}`;
-    //   const createResult = await todoHelperAdmin.create(resultSetName, todo);
-    //   expect(createResult.data[resultSetName].id).toBeDefined();
-    //   todo['id'] = createResult.data[resultSetName].id;
-
-    //   const todoUpdated = {
-    //     id: todo['id'],
-    //     content: 'Todo updated',
-    //     groupsField: [adminGroupName, devGroupName],
-    //   };
-    //   await expect(todoHelperNonAdmin.update(`update${modelName}`, todoUpdated)).rejects.toThrow(
-    //     'GraphQL error: Not Authorized to access updateTodoGroupFieldList on type Mutation',
-    //   );
-
-    //   await expect(async () => {
-    //     const getResult = await todoHelperNonAdmin.get({
-    //       id: todo['id'],
-    //     });
-    //   }).rejects.toThrowErrorMatchingInlineSnapshot(`"GraphQL error: Not Authorized to access getTodoGroupFieldList on type Query"`);
-
-    //   const listTodosResult = await todoHelperNonAdmin.list();
-    //   checkListItemExistence(listTodosResult, `list${modelName}s`, todo['id']);
-
-    //   await expect(
-    //     todoHelperNonAdmin.delete(`delete${modelName}`, {
-    //       id: todo['id'],
-    //     }),
-    //   ).rejects.toThrow('GraphQL error: Not Authorized to access deleteTodoGroupFieldList on type Mutation');
-    // });
-
-    // test('Admin user can give access to another group of users', async () => {
-    //   const modelName = 'TodoGroupFieldList';
-    //   const modelOperationHelpersAdmin = createModelOperationHelpers(appSyncClients[authProvider][userName1], schema);
-    //   const modelOperationHelpersNonAdmin = createModelOperationHelpers(appSyncClients[authProvider][userName2], schema);
-    //   const todoHelperAdmin = modelOperationHelpersAdmin[modelName];
-    //   const todoHelperNonAdmin = modelOperationHelpersNonAdmin[modelName];
-
-    //   const todo = {
-    //     content: 'Todo',
-    //     groupsField: [adminGroupName, devGroupName],
-    //   };
-    //   const resultSetName = `create${modelName}`;
-    //   const createResult = await todoHelperAdmin.create(resultSetName, todo);
-    //   expect(createResult.data[resultSetName].id).toBeDefined();
-    //   todo['id'] = createResult.data[resultSetName].id;
-
-    //   const todoUpdated = {
-    //     id: todo['id'],
-    //     content: 'Todo updated',
-    //     groupsField: [adminGroupName, devGroupName],
-    //   };
-    //   const updateResult = await todoHelperNonAdmin.update(`update${modelName}`, todoUpdated);
-    //   checkOperationResult(updateResult, todoUpdated, `update${modelName}`);
-
-    //   const getResult = await todoHelperNonAdmin.get({
-    //     id: todo['id'],
-    //   });
-    //   checkOperationResult(getResult, todoUpdated, `get${modelName}`);
-
-    //   const listTodosResult = await todoHelperNonAdmin.list();
-    //   checkListItemExistence(listTodosResult, `list${modelName}s`, todo['id'], true);
-
-    //   const deleteResult = await todoHelperNonAdmin.delete(`delete${modelName}`, {
-    //     id: todo['id'],
-    //   });
-    //   checkOperationResult(deleteResult, todoUpdated, `delete${modelName}`);
-    // });
-
-    // test('logged in user can perform custom operations', async () => {
-    //   const appSyncClient = appSyncClients[authProvider][userName2];
-    //   const todo = {
-    //     id: Date.now().toString(),
-    //     content: 'Todo',
-    //   };
-    //   const createTodoCustom = /* GraphQL */ `
-    //     mutation CreateTodoCustom($id: ID!, $content: String) {
-    //       addTodoPrivate(id: $id, content: $content) {
-    //         id
-    //         content
-    //       }
-    //     }
-    //   `;
-    //   const createResult = await appSyncClient.mutate({
-    //     mutation: gql(createTodoCustom),
-    //     fetchPolicy: 'no-cache',
-    //     variables: todo,
-    //   });
-    //   expect(createResult.data.addTodoPrivate).toBeDefined();
-
-    //   const getTodoCustom = /* GraphQL */ `
-    //     query GetTodoCustom($id: ID!) {
-    //       customGetTodoPrivate(id: $id) {
-    //         id
-    //         content
-    //       }
-    //     }
-    //   `;
-    //   const getResult = await appSyncClient.query({
-    //     query: gql(getTodoCustom),
-    //     fetchPolicy: 'no-cache',
-    //     variables: {
-    //       id: todo.id,
-    //     },
-    //   });
-    //   expect(getResult.data.customGetTodoPrivate).toHaveLength(1);
-    //   expect(getResult.data.customGetTodoPrivate[0].id).toEqual(todo.id);
-    //   expect(getResult.data.customGetTodoPrivate[0].content).toEqual(todo.content);
-    // });
-
-    // test('users in static group can perform custom operations', async () => {
-    //   const appSyncClient = appSyncClients[authProvider][userName1];
-    //   const todo = {
-    //     id: Date.now().toString(),
-    //     content: 'Todo',
-    //   };
-    //   const createTodoCustom = /* GraphQL */ `
-    //     mutation CreateTodoCustom($id: ID!, $content: String) {
-    //       addTodoStaticGroup(id: $id, content: $content) {
-    //         id
-    //         content
-    //       }
-    //     }
-    //   `;
-    //   const createResult = await appSyncClient.mutate({
-    //     mutation: gql(createTodoCustom),
-    //     fetchPolicy: 'no-cache',
-    //     variables: todo,
-    //   });
-    //   expect(createResult.data.addTodoStaticGroup).toBeDefined();
-
-    //   const getTodoCustom = /* GraphQL */ `
-    //     query GetTodoCustom($id: ID!) {
-    //       customGetTodoStaticGroup(id: $id) {
-    //         id
-    //         content
-    //       }
-    //     }
-    //   `;
-    //   const getResult = await appSyncClient.query({
-    //     query: gql(getTodoCustom),
-    //     fetchPolicy: 'no-cache',
-    //     variables: {
-    //       id: todo.id,
-    //     },
-    //   });
-    //   expect(getResult.data.customGetTodoStaticGroup).toHaveLength(1);
-    //   expect(getResult.data.customGetTodoStaticGroup[0].id).toEqual(todo.id);
-    //   expect(getResult.data.customGetTodoStaticGroup[0].content).toEqual(todo.content);
-    // });
-
-    // test('users not in static group cannot perform custom operations', async () => {
-    //   const appSyncClient = appSyncClients[authProvider][userName2];
-    //   const todo = {
-    //     id: Date.now().toString(),
-    //     content: 'Todo',
-    //   };
-    //   const createTodoCustom = /* GraphQL */ `
-    //     mutation CreateTodoCustom($id: ID!, $content: String) {
-    //       addTodoStaticGroup(id: $id, content: $content) {
-    //         id
-    //         content
-    //       }
-    //     }
-    //   `;
-    //   await expect(
-    //     appSyncClient.mutate({
-    //       mutation: gql(createTodoCustom),
-    //       fetchPolicy: 'no-cache',
-    //       variables: todo,
-    //     }),
-    //   ).rejects.toThrow('GraphQL error: Not Authorized to access addTodoStaticGroup on type Mutation');
-
-    //   const getTodoCustom = /* GraphQL */ `
-    //     query GetTodoCustom($id: ID!) {
-    //       customGetTodoStaticGroup(id: $id) {
-    //         id
-    //         content
-    //       }
-    //     }
-    //   `;
-    //   await expect(
-    //     appSyncClient.query({
-    //       query: gql(getTodoCustom),
-    //       fetchPolicy: 'no-cache',
-    //       variables: {
-    //         id: todo.id,
-    //       },
-    //     }),
-    //   ).rejects.toThrow('GraphQL error: Not Authorized to access customGetTodoStaticGroup on type Query');
-    // });
+    test('owner of a record can perform CRUD and subscription operations using default owner field', async () => {
+      await authTester.testRecordOwnerCrudOperationsWithDefaultOwnerField();
+    });
+
+    test('non-owner of a record cannot access or subscribe to it using default owner field', async () => {
+      await authTester.testNonRecordOwnerCannotAccessWithDefaultOwnerField();
+    });
+
+    test('custom owner field used to store owner information', async () => {
+      await authTester.testStoreOwnerInCustomField();
+    });
+
+    test('non-owner of a record cannot pretend to be an owner and gain access', async () => {
+      await authTester.testNonOwnerCannotPretendToBeOwner();
+    });
+
+    test('member in list of owners can perform CRUD and subscription operations', async () => {
+      await authTester.testListOwnersMemberCrudOperations();
+    });
+
+    test('non-owner of a record cannot add themself to owner list', async () => {
+      await authTester.testNonOwnerCannotAddThemselvesToList();
+    });
+
+    test('owner can add another user to the owner list', async () => {
+      await authTester.testOwnerCanAddAnotherUserToList();
+    });
+
+    test('users in static group can perform CRUD and subscription operations', async () => {
+      await authTester.testStaticGroupUserCrudOperations();
+    });
+
+    test('users not in static group cannot perform CRUD operations', async () => {
+      await authTester.testNonStaticGroupUserCrudOperations();
+    });
+
+    test('users in group stored as string can perform CRUD and subscription operations', async () => {
+      await authTester.testGroupUserCrudOperations();
+    });
+
+    test('users cannot spoof their group membership and gain access', async () => {
+      await authTester.testUsersCannotSpoofGroupMembership();
+    });
+
+    test('users in groups stored as list can perform CRUD and subscription operations', async () => {
+      await authTester.testListGroupUserCrudOperations();
+    });
+
+    test('users not part of allowed groups cannot access the records or modify allowed groups', async () => {
+      await authTester.testNotAllowedGroupUserCannotAccess();
+    });
+
+    test('Admin user can give access to another group of users', async () => {
+      await authTester.testAdminUserCanGiveAccessToAnotherUserGroup();
+    });
+
+    test('logged in user can perform custom operations', async () => {
+      await authTester.testLoggedInUserCustomOperations();
+    });
+
+    test('users in static group can perform custom operations', async () => {
+      await authTester.testStaticGroupUserCustomOperations();
+    });
+
+    test('users not in static group cannot perform custom operations', async () => {
+      await authTester.testNonStaticGroupUserCustomOperations();
+    });
   });
 };
