+++ packages/amplify-graphql-api-construct-tests/src/sql-tests-common/tests-sources/sql-dynamic-model-auth/sql-oidc-auth/stack-config.ts
@@ -1,12 +1,14 @@
 import { AUTH_TYPE } from 'aws-appsync';
 import { ImportedRDSType } from '@aws-amplify/graphql-transformer-core';
 import { schema as generateSchema } from './provider';
-import { StackConfig } from '../../../utils/sql-stack-config';
+import { StackConfig } from '../../../../utils/sql-stack-config';
+import { userGroupNames } from '../../../../utils/sql-provider-helper';
 
 export const stackConfig = (engine: ImportedRDSType): StackConfig => {
   const stackConfig: StackConfig = {
     schema: generateSchema(engine),
     authMode: AUTH_TYPE.OPENID_CONNECT,
+    userGroups: userGroupNames,
     oidcOptions: {
       triggers: {
         preTokenGeneration: `
