+++ packages/amplify-graphql-api-construct-tests/src/sql-tests-common/tests-sources/sql-dynamic-model-auth/sql-userpool-auth/provider.ts
@@ -1,5 +1,5 @@
 import { ImportedRDSType } from '@aws-amplify/graphql-transformer-core';
-import { convertToDBSpecificGraphQLString, generateDDL } from '../../../utils/sql-provider-helper';
+import { convertToDBSpecificGraphQLString, generateDDL, userGroupNames } from '../../../../utils/sql-provider-helper';
 
 export const schema = (engine: ImportedRDSType): string => /* GraphQL */ `
   input AMPLIFY {
@@ -30,7 +30,7 @@ export const schema = (engine: ImportedRDSType): string => /* GraphQL */ `
     authors: [String]
   }
 
-  type TodoStaticGroup @model @auth(rules: [{ allow: groups, groups: ["Admin"] }]) {
+  type TodoStaticGroup @model @auth(rules: [{ allow: groups, groups: ["${userGroupNames[0]}"] }]) {
     id: ID! @primaryKey
     content: String
   }
@@ -67,7 +67,7 @@ export const schema = (engine: ImportedRDSType): string => /* GraphQL */ `
     customGetTodoStaticGroup(id: ID!): [TodoNonModel] @sql(statement: "SELECT * FROM ${convertToDBSpecificGraphQLString(
       'TodoStaticGroup',
       engine,
-    )} WHERE id = :id") @auth(rules: [{ allow: groups, groups: ["Admin"] }])
+    )} WHERE id = :id") @auth(rules: [{ allow: groups, groups: ["${userGroupNames[0]}"] }])
   }
 
   type Mutation {
@@ -78,7 +78,7 @@ export const schema = (engine: ImportedRDSType): string => /* GraphQL */ `
     addTodoStaticGroup(id: ID!, content: String): TodoNonModel @sql(statement: "INSERT INTO ${convertToDBSpecificGraphQLString(
       'TodoStaticGroup',
       engine,
-    )} VALUES(:id, :content)") @auth(rules: [{ allow: groups, groups: ["Admin"] }])
+    )} VALUES(:id, :content)") @auth(rules: [{ allow: groups, groups: ["${userGroupNames[0]}"] }])
   }
 `;
